REMOVE: Ethereum Network Health Widget
