+++ launcher/src/components/UI/control-page/components/widgets/EpochSlot.vue
@@ -21,7 +21,7 @@
       />
       <ServiceLine
         label="INDEX"
-        :value="String(flag ? beaconControler : controlStore?.slotIndex + 1)"
+        :value="String(flag ? beaconControler : getSlotIndex())"
         :hover-text="
           controlStore.slotIndex < 0
             ? ''
@@ -47,24 +47,58 @@
 import { useSetups } from "@/store/setups";
 import { useControlStore } from "@/store/theControl";
 import { useFooter } from "@/store/theFooter";
-import { computed } from "vue";
+import { computed, onBeforeUnmount, onMounted, ref } from "vue";
 import NoData from "./NoData.vue";
 import ServiceLine from "../fragments/ServiceLine.vue"; // Assuming ServiceLine is the reusable component
 import i18n from "@/includes/i18n";
+import ControlService from "@/store/ControlService";
 
 const t = i18n.global.t;
 const controlStore = useControlStore();
 const footerStore = useFooter();
 const setupsStore = useSetups();
 
+const polling = ref(null);
+
 const beaconControler = computed(() => {
-  if (!controlStore.currentResult || controlStore.currentResult.beaconStatus === undefined) {
-    return "Checking Beacon Status...";
-  }
-  return controlStore.currentResult.beaconStatus !== 0 ? "No Running Beacon Node!" : "Loading...";
+  return "Loading...";
 });
+
+const refreshTimer = () => {
+  const intervalTime = setupsStore.selectedServicePairs?.network === "gnosis" ? 5000 : 11000;
+
+  polling.value = setInterval(() => {
+    currentEpochSlot();
+  }, intervalTime);
+};
+
+const currentEpochSlot = async () => {
+  try {
+    let res = await ControlService?.getCurrentEpochandSlot();
+    res.currentEpoch = res.current_epoch;
+    res.currentSlot = res.current_slot;
+    delete res.current_epoch;
+    delete res.current_slot;
+    controlStore.currentResult = res;
+  } catch (error) {
+    console.error("An error occurred while fetching currentEpochandSlot:", error);
+  }
+};
+
+onMounted(() => {
+  refreshTimer();
+});
+
+onBeforeUnmount(() => {
+  clearInterval(polling.value);
+});
+
+const getSlotIndex = () => {
+  const slot = controlStore?.currentResult?.currentSlot;
+  return slot % controlStore.currentResult.slotsPerEpoch;
+};
 
 const isConsensusMissing = computed(() => footerStore?.missingServices?.includes("consensus"));
 
-const flag = computed(() => !controlStore?.currentResult || controlStore?.currentResult?.beaconStatus !== 0);
+const flag = computed(() => !controlStore?.currentResult);
 </script>
